requirements.txt
absl-py==1.4.0
action-msgs==1.2.1
action-tutorials-interfaces==0.20.3
action-tutorials-py==0.20.3
actionlib-msgs==4.2.3
addict==2.4.0
ament-cmake-test==1.3.6
ament-copyright==0.12.9
ament-cppcheck==0.12.9
ament-cpplint==0.12.9
ament-flake8==0.12.9
ament-index-python==1.4.0
ament-lint==0.12.9
ament-lint-cmake==0.12.9
ament-package==0.14.0
ament-pep257==0.12.9
ament-uncrustify==0.12.9
ament-xmllint==0.12.9
angles==1.15.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
argcomplete==1.8.1
attrs==21.2.0
Automat==20.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
builtin-interfaces==1.2.1
catkin-pkg-modules==1.0.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cmake==3.27.2
colorama==0.4.4
command-not-found==0.3
composition-interfaces==1.2.1
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.0
cryptography==3.4.8
cupshelpers==1.0
cv-bridge==3.2.1
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
deluge==2.0.3
demo-nodes-py==0.20.3
diagnostic-msgs==4.2.3
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.17.1
domain-coordinator==0.10.0
empy==3.3.4
example-interfaces==0.9.3
examples-rclpy-executors==0.15.1
examples-rclpy-minimal-action-client==0.15.1
examples-rclpy-minimal-action-server==0.15.1
examples-rclpy-minimal-client==0.15.1
examples-rclpy-minimal-publisher==0.15.1
examples-rclpy-minimal-service==0.15.1
examples-rclpy-minimal-subscriber==0.15.1
exceptiongroup==1.2.0
eyeD3==0.8.10
filelock==3.6.0
Fiona==1.9.4.post1
flake8==4.0.1
flatbuffers==23.5.26
fonttools==4.42.1
fs==2.4.12
ftfy==6.1.1
fuse-python==1.0.2
fvcore==0.1.5.post20221221
gast==0.5.2
GeoIP==1.3.2
geometry-msgs==4.2.3
geopandas==0.13.2
greenlet==1.1.2
grpcio==1.30.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.20.2
httpx==0.26.0
hyperlink==21.0.0
idna==3.3
ifaddr==0.1.7
image-geometry==3.2.1
imageio==2.31.3
IMDbPY==2021.4.18
importlib-metadata==6.8.0
incremental==21.3.0
iniconfig==1.1.1
interactive-markers==2.3.2
iopath==0.1.10
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
kazam==1.4.5
keyring==23.5.0
kiwisolver==1.4.5
lark==1.1.1
laser-geometry==2.4.0
launch==1.0.4
launch-ros==0.19.6
launch-testing==1.0.4
launch-testing-ros==0.19.6
launch-xml==1.0.4
launch-yaml==1.0.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
lifecycle-msgs==1.2.1
lit==16.0.6
logging-demo==0.20.3
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
map-msgs==2.1.0
MarkupSafe==2.0.1
matplotlib==3.7.2
mccabe==0.6.1
mediapipe==0.10.3
message-filters==4.3.3
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==1.3.0
msgpack==1.0.3
mypy==1.9.0
mypy-extensions==1.0.0
nav-msgs==4.2.3
nemo-emblems==5.6.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.1
numpy==1.25.2
oauthlib==3.2.0
omegaconf==2.3.0
onboard==1.4.1
openai==1.7.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
osrf-pycommon==2.0.2
packaging==21.3
PAM==0.4.2
pandas==2.0.3
pcl-msgs==1.0.0
pendulum-msgs==0.20.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.10.0
pluggy==0.13.0
ply==3.11
portalocker==2.7.0
prettytable==3.6.0
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
pycurl==7.44.1
pydantic==2.5.3
pydantic_core==2.14.6
pydocstyle==6.1.1
pydot==1.4.2
pyelftools==0.27
pyflakes==2.4.0
pygame==2.1.2
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pylibacl==0.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.11.7
pyproj==3.6.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==6.2.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-gnupg==0.4.8
python-magic==0.4.24
python-qt-binding==1.1.2
python-xlib==0.29
pythran==0.10.0
pytz==2022.1
PyWavelets==1.4.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
qt-dotgraph==2.2.2
qt-gui==2.2.2
qt-gui-cpp==2.2.2
qt-gui-py-common==2.2.2
quality-of-service-demo-py==0.20.3
rcl-interfaces==1.2.1
rclpy==3.3.11
rcutils==5.1.4
rencode==1.0.6
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
resource-retriever==3.1.1
rmw-dds-common==1.6.0
roman==3.3
ros2action==0.18.7
ros2bag==0.15.8
ros2cli==0.18.7
ros2component==0.18.7
ros2doctor==0.18.7
ros2interface==0.18.7
ros2launch==0.19.6
ros2lifecycle==0.18.7
ros2multicast==0.18.7
ros2node==0.18.7
ros2param==0.18.7
ros2pkg==0.18.7
ros2run==0.18.7
ros2service==0.18.7
ros2topic==0.18.7
rosbag2-interfaces==0.15.8
rosbag2-py==0.15.8
rosdistro-modules==0.9.0
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rospkg-modules==1.5.0
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.4
rqt-bag-plugins==1.1.4
rqt-console==2.0.2
rqt-graph==1.3.0
rqt-gui==1.1.5
rqt-gui-py==1.1.5
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.5
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scikit-image==0.21.0
scipy==1.11.2
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sensor-msgs==4.2.3
sensor-msgs-py==4.2.3
service-identity==18.1.0
setproctitle==1.2.2
shape-msgs==4.2.3
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sounddevice==0.4.6
soupsieve==2.3.1
sros2==0.10.4
statistics-msgs==1.2.1
std-msgs==4.2.3
std-srvs==4.2.3
stereo-msgs==4.2.3
sympy==1.12
systemd-python==234
tabulate==0.9.0
teleop-twist-keyboard==2.3.2
termcolor==2.3.0
texttable==1.6.4
tf2-geometry-msgs==0.25.5
tf2-kdl==0.25.5
tf2-msgs==0.25.5
tf2-py==0.25.5
tf2-ros-py==0.25.5
tf2-tools==0.25.5
tifffile==2023.8.30
timm==0.6.7
tinycss2==1.1.1
tldextract==3.1.2
toml==0.10.2
tomli==2.0.1
topic-monitor==0.20.3
torch==2.0.1+cu117
torchvision==0.15.2+cu117
tornado==6.1
tqdm==4.66.1
trajectory-msgs==4.2.3
triton==2.0.0
turtlesim==1.4.2
Twisted==22.1.0
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
Unidecode==1.3.3
unique-identifier-msgs==2.2.1
urllib3==1.26.5
virtualenv==20.13.0+ds
visualization-msgs==4.2.3
wadllib==1.3.6
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.2.3
xdg==5
xkit==0.0.0
xlrd==1.2.0
yacs==0.1.8
yapf==0.40.1
youtube-dl==2021.12.17
zipp==1.0.0
zope.interface==5.4.0
